Revert "refactor: doorコマンドの通信をアプリケーションポートに抽象化 (#200)"
This reverts commit a017418fee9deaef2ca6f9e19557370604444227.
