+++ src/application/ports/deps.ts
@@ -2,7 +2,6 @@ import type { DiscordChannelPort } from "./discordChannelPort";
 import type { DiscordGuildPort } from "./discordGuildPort";
 import type { DiscordMemberPort } from "./discordMemberPort";
 import type { DiscordMessagePort } from "./discordMessagePort";
-import type { DoorStatusPort } from "./doorStatusPort";
 import type { EmailAuthPort } from "./emailAuthPort";
 import type { ITSCorePort } from "./itsCorePort";
 import type { ScheduledMessagePort } from "./scheduledMessagePort";
@@ -19,5 +18,4 @@ export interface AppDeps {
   discordGuildPort: DiscordGuildPort;
   discordMessagePort: DiscordMessagePort;
   scheduledMessagePort: ScheduledMessagePort;
-  doorStatusPort: DoorStatusPort;
 }
