--- src/infrastructure/door-status/index.ts
@@ -1,1 +1,0 @@
-export { WebSocketDoorStatusAdapter } from "./webSocketDoorStatusAdapter";
