+++ src/main.ts
@@ -3,7 +3,6 @@ import { initializeScheduledMessagesFromConfig } from "@application/usecases";
 import { loadConfig } from "@config/environment";
 import { CustomClient } from "@domain/types";
 import { DiscordServerAdapter } from "@infrastructure/discordjs";
-import { WebSocketDoorStatusAdapter } from "@infrastructure/door-status";
 import { FirebaseEmailAuthAdapter } from "@infrastructure/firebase";
 import { ITSCoreAdaptor } from "@infrastructure/itscore";
 import logger from "@infrastructure/logger";
@@ -37,8 +36,6 @@ async function main() {
 
     // Composition Root: adapter を生成し依存オブジェクトを組み立てる
     const discordServerAdapter = new DiscordServerAdapter(client);
-    const doorStatusUrl =
-      process.env.DOOR_STATUS_WS_URL ?? "wss://its-status.woody1227.com/";
     const deps: AppDeps = {
       itsCorePort: new ITSCoreAdaptor(),
       emailAuthPort: new FirebaseEmailAuthAdapter(),
@@ -47,7 +44,6 @@ async function main() {
       discordGuildPort: discordServerAdapter,
       discordMessagePort: discordServerAdapter,
       scheduledMessagePort: memoryScheduledMessageRepository,
-      doorStatusPort: new WebSocketDoorStatusAdapter(doorStatusUrl),
     };
 
     // Cron manager に依存オブジェクトを設定
